Add sa::parsing_cntx define_ident test -- Failing
There is a bug that should be fixed in the sa::define_identifier()
function and not only there, but the whole
sa::parsing_context::m_all_scopes, too. It is based on a vector of two parts:
- std::string name;
- identifier.

However the identifier holds its own name, so the vector does not need
to hold it for the ident.